src/test/java/com/neobank/web/OrderControllerTest.java
package com.neobank.web;

import com.fasterxml.jackson.databind.ObjectMapper;
import com.neobank.domain.interfaces.OrderService;
import com.neobank.dto.Price;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.RequestBuilder;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;

import java.util.Arrays;

import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@RunWith (SpringRunner.class)
@WebMvcTest(OrderController.class)
public class OrderControllerTest {
    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private OrderService orderService;

    @Test
    public void shouldReturnOkStatus() throws Exception{
        RequestBuilder requestBuilder= MockMvcRequestBuilders.post ("/checkout")
                .content(asJsonString(Arrays.asList ("001", "001", "002")))
                .contentType(MediaType.APPLICATION_JSON)
                .accept(MediaType.APPLICATION_JSON);
        mockMvc.perform(requestBuilder).andExpect(status().isOk());
    }

    @Test
    public void shouldGetCheckOutCartValue() throws Exception {
        when(orderService.checkOut (Arrays.asList ("001", "002"))).thenReturn(new Price (100.0));
        RequestBuilder requestBuilder = MockMvcRequestBuilders.post("/checkout")
                .content(asJsonString(Arrays.asList ("001", "002")))
                .contentType(MediaType.APPLICATION_JSON)
                .accept(MediaType.APPLICATION_JSON);
        mockMvc.perform(requestBuilder).andExpect(status().isOk()).andExpect(content().json("{\"price\":100.0}"));
    }

    private  static String asJsonString(final Object obj) {
        try {
            final ObjectMapper mapper = new ObjectMapper();
            final String jsonContent = mapper.writeValueAsString(obj);
            return jsonContent;
        } catch (Exception e) {
            throw new RuntimeException(e);
        }
    }
}
